+++ src/Connection/Response/ErrorResponse.php
@@ -1,0 +1,15 @@
+<?php
+namespace Disque\Connection\Response;
+
+class ErrorResponse extends BaseResponse
+{
+    /**
+     * Parse response
+     *
+     * @return ResponseException Response
+     */
+    public function parse()
+    {
+        return new ResponseException((string) $this->data);
+    }
+}