+++ src/Connection/Response/ResponseException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Disque\Connection\Response;
+
+use Disque\Connection\ConnectionException;
+
+class ResponseException extends ConnectionException
+{
+}