+++ src/Connection/Response/StringResponse.php
@@ -1,0 +1,15 @@
+<?php
+namespace Disque\Connection\Response;
+
+class StringResponse extends BaseResponse
+{
+    /**
+     * Parse response
+     *
+     * @return string Response
+     */
+    public function parse()
+    {
+        return (string) $this->data;
+    }
+}