+++ src/Connection/ResponseException.php
@@ -1,6 +1,0 @@
-<?php
-namespace Disque\Connection;
-
-class ResponseException extends ConnectionException
-{
-}