+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All Notable changes will be documented in this file. This project adheres to 
 [Semantic Versioning](http://semver.org/).
 
+## [1.3.1]
+
+### Changed
+- Exception `Disque\Connection\ResponseException` has been moved to 
+`Disque\Connection\Response\ResponseException`
+
 ## [1.3.0] - 2015-05-18
 
 ### Added